+++ src/mruby_integration/models/image.cpp
@@ -60,6 +60,19 @@ mrb_Image_initialize(mrb_state* mrb, mrb_value self) -> mrb_value
 }
 
 auto
+mrb_Image_unload(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  Image* image;
+
+  Data_Get_Struct(mrb, self, &Image_type, image);
+  mrb_assert(image != nullptr);
+
+  UnloadImage(*image);
+
+  return mrb_nil_value();
+}
+
+auto
 mrb_Image_get_data(mrb_state* mrb, mrb_value self) -> mrb_value
 {
   Image* image;
@@ -97,6 +110,8 @@ append_models_Image(mrb_state* mrb)
   MRB_SET_INSTANCE_TT(Image_class, MRB_TT_DATA);
   mrb_define_method(
     mrb, Image_class, "initialize", mrb_Image_initialize, MRB_ARGS_REQ(5));
+  mrb_define_method(
+    mrb, Image_class, "unload", mrb_Image_unload, MRB_ARGS_NONE());
   mrb_define_method(
     mrb, Image_class, "data", mrb_Image_get_data, MRB_ARGS_NONE());